Fix error in sefran with empty last day
The sefran displays the following error in the log when the last day in
the sefran ROOT directory has no minute file:
 sefran3.pl: Month '-1' out of range 0..11 at /opt/webobs/WebObs-2.1.0/CODE/cgi-bin/sefran3.pl line 357.

This error also appears on an error web page that replaces the expected
empty sefran when displaying /cgi-bin/sefran3.pl. (This behaviour seems
to have appeared sometime between WebObs versions 1.8.0 and 2.1.0).

This little commit prevents the error and allows the expected display of
the empty sefran.