Fix false negatives by cmpsql
'cmpsql' is used at the end of 'setup' to compare the schema of existing
sqlite databases with the schema of a newly created temporary database.

The script could however return false negatives for 2 reasons:
  * any existing sqlite_sequence system table in the existing db (this
  kind of table cannot be removed, even if unused)
  * for an unknown reason, some database table schemas are displayed
  this way: "CREATE TABLE jobs (" while others are displayed "CREATE
  TABLE jobs(" (with the same issue on the closing parenthesis).

These issues are hopefully fixed in this commit, and in case of any
discrepancy between the schemas, a warning is printed, together with the
full schemas of the two databases so the admin can instantly look for
the difference.