Fix annoying display bug on Chrome/Chromium browsers
Although WebObs officially only supports Firefox, this commit fixes a
potentially annoying/"dangerous" bug under Chrome/Chromium (occasionally
used by some users) : all events in the Gazette were shifted by 1 day to
the right.

This was a consequence of Firefox not supporting the 'rowspan=0'
behaviour in compatibility mode (aka "quirks mode") used for Webobs
pages. The 'rowspan=0' attribute was used in the table header cell
displaying the week number (rotated by 270° using CSS) to span the cell
on the whole column, but, as it was not interpreted by Firefox, table
rows were shifted by 1 column/day using empty 'td' cells to account for
this header cell. Under Chrome, both the 'rowspan=0' and the empty 'td'
cells were rendered, resulting in an extra shift to the right.

Simply removing 'rowspan=0' fixes the problem under Chrome/Chromium,
with no consequence under Firefox as it is not handled by this
navigator on this page.