Link system-wide perl modules to current WebObs version
The 'setup' script used to link the WebObs perl modules to the directory
/opt/webobs/WebObs-X.Y.Z/CODE/cgi-bin/, which could be an issue after an
upgrade if the administrator forgot to confirm the update of these
links.

As the modules must point to the WebObs version in use, it makes more
sense to point to the /opt/webobs/CODE/ link that decides the WebObs
version currently in use.