Add an expert mode to the 'setup' script
The 'setup' script makes many assumptions about the user's environment
that may not always be true.

This commit introduces an "expert mode" that will let the user proceed
with the rest of the install/upgrade even if the environment does not
meet all the requirements of 'setup'. (A warning is printed and a
confirmation requested before continuing.)

This gives users the liberty to use a different environment but still
allows to perform upgrades or installations through this script as long
as they know what they are doing.