Fixes some security vulnerabilities in a few CGI scripts
Theses changes fix important security vulnerabilities in the CGI scripts
handling the GRIDS.

 * implements the WebObs::Utils::checkParam function that will check a value
   against a strict regular expression to avoid commands injections in the
   code.
 * adds checks of CGI parameters using WebObs::Utils::checkParam in
   cgi-bin/*GRID*.pl that needed it the most.
 * for security and optimisation, replaces calls to unix commands by equivalent
   perl functions and use DBI for SQLite requests instead of the "sqlite3"
   system command.
 * adds many defined() tests to clarify and avoid warning floods to
   stderr/logs.
 * in some scripts, reorganises control flow to ease debugging and maintenance.

Note that this calls for testing, as the parameter checks only allow the strict
necessary subset of values and may fail in some case if values contain
unexpected characters. In some cases, the regular expressoin passed to
checkParam should be (very carefully) adapted.