Reorganise and slightly strengthen register.pl
Theses changes aims at strengthening the sensitive script 'register.pl'
that is not protected by any password.

 * reorganise the control flow in the script to ease maintenance.
 * add even more controls of the CGI parameter values provided by the
   visitor.
 * introduce a REGISTRATION_LOGFILE configuration variable in WEBOBS.rc
   to designate the registration log file (which defaults to the
   value previously used, ${ROOT_DATA_DB}/reglog).

Theses changes introduces some limitations on the values the user can
enter in the form. UTF-8 letters are supported as well as some
legimitate special characters, so this should hopefully not be a problem
for the fullname field.

Althought care has been taken not to get in the way for legitimate uses,
some tweaks might be required for the accepted values in the form.
Further tests and first uses in real conditions should help us with
these tweaks.