Better handling of 'lastevent' to display the latest b3
When using the 'lastevent' value to show the B3 of the latest event,
showOUTG.pl was not displaying the date correctly on the page.

This commit slightly change the way showOUTG.pl handles this 'lastevent'
case.

As a reminder, using 'g=lastevent' will still show all the B3 reports of
the latest event (or the B3 itself if there is only one, as usual).
  e.g.: /cgi-bin/showOUTG.pl?grid=PROC.MY_B3_PROC&ts=events&g=lastevent
and using 'g=lastevent/b3' will force the correct display of the very
last B3 :
  e.g.: /cgi-bin/showOUTG.pl?grid=PROC.MY_B3_PROC&ts=events&g=lastevent/b3
(this was previously possible using 'g=lastevent////b3', which still works)