Improve error handling in postCLB.pl
Changed readCfg to have it return an empty array, hash, or undef scalar
if called with and undef argument, or when reading an empty file.

This allows postCLB.pl to now correctly handle readCfg return value, but
we had to exit the script in case of error to avoid emptying the
calibration file.