Fix event photo src and href in WebObs::Events
The WebObs::Events::photoStrip subroutine was using '../NODES' as a URI
in src and href, therefore breaking image display and link in the photos
of node events (at least if URN_NODES is not /NODES).

The same issue was recently fixed in formUPLOAD.pl by 4d0eebd (related
to issue #28).