Sort domain and form lists in Grid edition form
The drop-down list of Grid domains as well as the list of forms in the
Grid edition form are currently unsorted and present their elements in a
different order each time one uses the page.

For better consistency of element orders, we now sort these lists while
building the corresponding HTML 'select' elements.