Fix and improve autoregistration of new user
As the issue #33 showed, in case an UID is already in use in the system,
the registration request of a new user will fail (if autoregistration is
enabled), as the failed database insert statement will raise an error
and won't let the code handle the error and try a new UID.

Additionally, it was discovered that the alternative UID to an existing
`XB` UID would have the form `XB2`, `XB23`, `XB234`, etc. instead of the
expected `XB2`, `XB3`, `XB4`, etc.

This commit fixes both of these bugs and improve the autoregistration
process by first checking in the database whether an UID already exists
before trying to insert the new user. This is an improvement, as we were
previously assuming that the only possible cause of a failed `insert`
statement was an already existing UID, and things would go wrong with
another database issue.

This closes #33.