Fix channels sorting in calibration form
When using more than 100 channels of the same date and time, the
channel lines were wrongly worted because of the default string
comparison operator.

We now explicitely sort on date and time (using a string comparison),
and use a numerical comparison to sort on the channel number as a
secondary sorting column.