Fix listing of FID_* variables on node page
FID_* node variables are presently listed in a different order each time
one loads a node page.

With this commit, the list of these variables is now sorted before
display, so that it always appears in the same order.  (It also slightly
improve code readibility for the making of this list.)