A simple testing framework for the DSV preprocessor
Add `run_all_tests.sh`, a simple script to run tests on the DSV
preprocessor script (like `dsv_generic`), along with a few tests
defined in the test_*.sh files.

This is only intended as a tool to avoid regressions in the `dsv_generic`
and clarify some corner cases, and should mainly be useful to WebObs
author and contributors.

See the README.md file for information about how to run these tests.