Add advice to SC3 FDSNWS users about not locatable events
In the default MC3 configuratin file, add a comment on the limitation of
the SC3 FDSNWS XML schema that maps the "not locatable" event types to
"other events".

When using the FDSN web services, this latter value should therefore be
included in SC3_EVENT_TYPES_NOLOCATION to properly hide location of
events classified as "not locatable" in SeisComp3.

When using the SC3 XML files (generated by the SC3 server), this value
is not necessary and might potentially lead to unwanted hidden
location, though in some very rare cases.  It was therefore not added to
the provided default value of SC3_EVENT_TYPES_NOLOCATION.