Allow to remove and erase data in the GAZ form
This commit enables the possibility to remove a line of data in the form
GAZ (just like the EAUX form already allows).

postGAZ.pl already had this functionnality ready but it was not enabled
in formGAZ.pl and showGAS.pl.

Also introduce a few small readibility improvements and fix some
undefined variables errors in formGAZ.pl.