Introduce the 'hideloc' CGI parameter of sefran3.pl
The new 'hideloc' CGI parameter of sefran3.pl is similar to the
'hideloc' parameter of mc3.pl and hides locations of events in the
popovers in the hourly thumbnail view.

If not present, it defaults to the inverted value of
MC3_EVENT_DISPLAY_LOC from the Sefran configuration.

This is especially useful for kiosk mode displays of the Sefran, to
avoid unnecessarily overloading the servers.