Fix trashed event preference when displaying sefran hour
Currently, when an admin user browses the Sefran hour thumbnails and
chooses not to display the trashed events, these events will still
appear when viewing the hour full view after clicking on a sefran
thumbnail.

This commit applies to the hour full view the 'trash' setting chosen by
the user in the thumbnail view.