Add seismic energy calculation to the MC
This commit implements the seismic energy calculation to the Main
Courante.
- event's energy is shown in a popover in the Magnitude column of the
  MC whenever the magnitude is available.
- the total energy of the filtered events is displayed on the top, below
  the total number of events.
- two new charts are available :
  * a graph showing the energy of each type of event, cumulated by day.
  * a graph of the total energy of all types of events, cumulated by
    day. (I was asked to make this one a simple unfilled line.)

The color of the total energy graph is hardcoded in the script (in blue,
but any color other than black is welcome), but can be customised by
adding an 'ENERGY' type in the MC3_Codes.conf file corresponding to the
MC. (Although it does not make much sense to define such a type of
event, it was an easy way to allow for customisation. Improved
customisation means are most welcome.)