New scheduler option to merge jobs stderr and stdout
A new boolean configuration option `MERGE_JOG_LOGS` in the scheduler
configuration allows to write the job stdout and stderr outputs to a
same log file (using prefix `.log` instead of `.stdout`/`.stderr`).

If `MERGE_JOG_LOGS` is not `y`, `yes` or `1` or is missing, the old
behaviour is followed, and outputs will be written to different files.