Improve scheduler commands outputs
Now sort the list of jobs printed by the scheduler on commands like
'runq', 'jobq' and 'qs'.

In case of 'runq', both the list of jobs and the provided information
are sorted to ensure a constant order. The jobs starting date and time
are also now printed in human readable format to help the administrator.