setup now installs or updates the systemd services
On systems using systemd, the setup script now proposes to automatically
install or update the wopostboard.service and woscheduler.service files
by adapting the provided templates to the system environment (it changes
the DOCROOT directory and webobs user and group name).

The manual and release notes have be updated to reflect this new
feature.