Bug and security fixes and improved HTML interface
These changes fix a bug in the Scheduler Runs web page that prevents
from displaying the link that allows to kill running jobs. Fix #66.

Also, in the same HTML table, the "Elapsed" column now shows "still
running" for unfinished jobs instead of a negative time.

This commit however mainly replaces calls to qx() in schedulerMgr.pl and
schedulerRuns.pl by calls to the DBI interface for running SQLite
queries and by calls to the function
WebObs::Scheduler::scheduler_client() to contact the scheduler.

This last change fixes a security issue in both scripts allowing to run
arbritrary commands on the server to users with the admin level.