Fix redirection problem in sefran replay mode
The issue happens when using multiple SefraN.
When the user identifies seismic events in replay mode on the non-default SefraN, he is redirected to the default SefraN after validating the first event.
Whith this fix, the user will be redirected to the SefraN he came from.
The fix uses the URL parameters mc3 and s3.