Fix logo URL in register.pl
The LOGO_DEFAULT value from WEBOBS.rc is used in default proc
configurations as the path to a default logo file.

The register.pl script was using the LOGO_DEFAULT value from WEBOBS.rc
as the URL to the logo to display, which is incompatible with other uses
of LOGO_DEFAULT in default proc configurations, where it is expected to
be the path to a default logo file.

This fix introduces the URN_LOGO_DEFAULT in WEBOBS.rc as the URL to a
default logo and uses it in register.pl.