Fix the scheduler timeline for timezones east of Greenwich
The schedulerRuns.pl script was systematically converting the unix
epoch representing the end date of a run  and was relying on null values
being converted to the day '1970-01-01'.  The convertion was however
also converted to localtime, resulting in a day of '1969-12-31' for
timezones east of Greenwich.

This was preventing the running jobs from showing up in the timeline of
job runs. It was fixed using a saner approach for the handling of the
date format conversion.

This commit also introduces more explicit variable names in modified
code blocks and fixes bugs introduced in recent changes of the script
that were wrongly excluding running jobs from the timeline.

A quick change to scheduler.pl fixes the format of the log filename in
the 'runs' database table (also displayed in the 'Logs' column of the
HTML table of runs).