Fix several issues in calibration file edition
Now ignore the INSTALL_DATE from the node configuration if it is not a
valid date (it could potentially freeze the browser). This fixes the
issue #76.

Also fix the change of the submit button label that never worked as
intended (not sure it was worth fixing as the whole thing is confusing
as hell, but well).